Change get shipments by permission date

+++ Logitude/Logitude.BL/GlobalModel/EntityPMs/TenantManagementPM.cs
@@ -383,7 +383,7 @@ namespace Logitude.BL.GlobalModel.EntityPMs
         public int LastWeekCreatedTariffs { get; set; }
         [CustomValidation(typeof(Validators.ValidationClass), "ValidateClass")]
         public int LastMonthCreatedTariffs { get; set; }
-        public int? PermissionBuildMonths { get; set; }
+        public double? PermissionBuildMonths { get; set; }
 
         public int ScheduledTasksLimitPerReport { get; set; }
         public bool IsHybrid { get; set; }
